// client/emerald-app/src/app/users/users.ts
// export class User {
//     id?: string;
//     username?: string;
//     password?: string;
//     firstName?: string;
//     lastName?: string;
//     token?: string;
// }

export class User {
    id?: string;
    name?: string;
    password?: string;
    admin?: string;   
}

export class UserInfo {
    id?: string;
    name?: string;
    password?: string;
    admin?: string;   
  }